Rename notifier_host::dbus to more descriptive notifier_host::proxy
