+++ crates/notifier_host/src/proxy/dbus_status_notifier_item.xml
